+++ back-end/src/main/java/com/br/coworkingSystem/repositories/ProdutoServicoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.br.coworkingSystem.model.ProdutoServico;
 
 @Repository
-public interface ProdutoServicoRepository extends JpaRepository<ProdutoServico, Integer> {
+public interface ProdutoServicoRepository extends JpaRepository<ProdutoServico, Long> {
 
 }
